allow to compile from plain sources
The master branch allows for specifying the icon and applications directory via a CMake parameter.
Use these to avoid having to patch the build system files.

Use the patch utility to patch marble.desktop just after "make install".
This approach should be safe in terms of preserving permissions of the marble.desktop file.